move errors-related stuff in app/token into errors.go and errors_test.go
